Rename file so that it is different from module name
