+++ rust/chains/hyperlane-cosmos/src/mailbox.rs
@@ -1,7 +1,7 @@
+pub use contract::CosmosMailbox;
 pub use delivery_indexer::CosmosMailboxDeliveryIndexer;
 pub use dispatch_indexer::CosmosMailboxDispatchIndexer;
-pub use mailbox::CosmosMailbox;
 
+mod contract;
 mod delivery_indexer;
 mod dispatch_indexer;
-mod mailbox;
